# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MEDX - Medical Education and Development Exchange API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [contact]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"
...
components:
  schemas:
...
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        subject:
          type: ["string", "null"]
        message:
          type: string
    ContactSubmission:
      type: object
      required: [id, name, email, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        name:
          type: ["string", "null"]
    NewsletterSubscription:
      type: object
      required: [id, email, subscribedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        subscribedAt:
          type: string
          format: date-time
...