# packages/api-zod/src/generated/types/newsletterSubscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MEDX - Medical Education and Development Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscription {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  subscribedAt: Date;
}